Modified DG to cost money.
Messy as f.

+++ pappab0t/Modules/Inventory/Inventory.cs
@@ -6,7 +6,7 @@ namespace pappab0t.Modules.Inventory
     {
         public string Id { get; set; }
         public string UserId { get; set; }
-        public int BEK { get; set; }
+        public decimal BEK { get; set; }
         public List<Item> Items { get; set; }
 
         public Inventory()
